requirements.txt
asyncpg==0.30.0
hiredis==3.1.0
jieba==0.42.1
ldap3==2.9.1
llama-index-core==0.12.28
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
nltk==3.9.1
llama-index-embeddings-ollama==0.6.0
llama-index-llms-ollama==0.5.4
llama-index-storage-docstore-redis==0.3.0
llama-index-storage-index-store-redis==0.4.0
llama-index-vector-stores-redis==0.5.0
RapidFuzz==3.13.0
streamlit-aggrid==1.1.3.post1
streamlit-searchbox==0.1.22
thulac==0.2.2
tinycss2==1.4.0
